Upload functional design
The current design has passed the verification.

All the classes (Generator, Driver, Monitor, etc) are written in the
same file (SimClasses.sv). This is not correct and will be split into
separated files (Driver.sv, Monitor.sv, etc).

The Driver and the Monitor are communicating using events as a
handshake. This should be avoid since this make the Testbench less
reusable